Make graphql a real dep and check-in lib/ (#1)
## Summary:
We've upgrade graphql in webapp/services/static/ to be a version that is incompatible with this eslint plugin.  In order for this eslint plugin to get a version of graphql that it's compatible with, we need to make graphql a 'dep' instead of a 'peerDep'.  This will for npm/yarn to install a separate version of 'graphql' just for this plugin.

To avoid having to publish this plugin, I've checked-in the lib/ folder.

Issue: FEI-5233

## Test plan:
- update https://github.com/Khan/webapp/pull/15741 to point to this repo/commit

Author: kevinbarabash

Reviewers: kevinbarabash, jeresig

Required Reviewers:

Approved By: jeresig

Checks:

Pull Request URL: https://github.com/Khan/eslint-plugin-graphql/pull/1